+++ projects/mk-kit/dnd/index.ts
@@ -2,6 +2,7 @@ export * from './drag-drop.types';
 export * from './drag-drop-utils';
 export * from './drag-drop-registry';
 export * from './drop-list';
+export * from './drop-zone';
 export * from './drag';
 export * from './drag-handle';
 export * from './sortable-list';